athena query

+++ athena/get-athena-queries.sh
@@ -21,7 +21,7 @@ c2=1
 if [ "$count" -gt "0" ]; then
     count=`expr $count - 1`
     #loop through query id's
-    for i in `seq 0 20`; do
+    for i in `seq 0 $count`; do
         qid=`echo $awsout | jq ".${pref[(${c})]}[(${i})]" | tr -d '"'`
         #echo "quid=$qid"
         cm="${cmd[$c2]} $qid"
@@ -39,6 +39,9 @@ if [ "$count" -gt "0" ]; then
         printf "resource \"%s\" \"%s\" {" $ttft $cname > $cname.tf
         printf "}" $cname >> $cname.tf
         terraform import $ttft.$cname $qid
+        if [ $? -ne 0 ]; then
+            exit
+        fi
         terraform state show $ttft.$cname > t2.txt
         rm $cname.tf
 
@@ -55,7 +58,7 @@ if [ "$count" -gt "0" ]; then
             t1=`echo "$line"` 
             if [[ ${t1} == *"="* ]];then
                 tt1=`echo "$line" | cut -f1 -d'=' | tr -d ' '` 
-                tt2=`echo "$line" | cut -f2- -d'=' | tr -d ' '`  
+                tt2=`echo "$line" | cut -f2- -d'='`  
                 if [[ ${tt1} == "arn" ]];then	
                 	#printf "acl = \"private\" \n" >> $fn
                     #printf "force_destroy = false \n" >> $fn
@@ -63,12 +66,18 @@ if [ "$count" -gt "0" ]; then
                     skip=1
                 fi
                 if [[ ${tt1} == "query" ]];then
-                    echo "tt2="
-                    echo $tt2
-                    tt3=`echo $tt2 | sed 's/"/\\"/g'`
-                  
-                    echo $tt3
+                    #echo "tt2="
+                    #echo $tt2
+                    echo query=$tt2 | sed 's/",/\\",/g' | sed 's/,"/,\\"/g' > t5.txt
+                    tt5=`echo $tt2 | sed 's/",/\\",/g' | sed 's/,"/,\\"/g'`
+                    #echo "part1="$tt1
+                    #echo "part2="$tt5
                     
+                    t1=`printf "%s=%s" "$tt1" "$tt5"`
+                    t1=`cat t5.txt`
+                    #echo "query="
+                    #echo $t1
+                             
                 fi
